単語判定をチェックポイント法に

+++ Source/Voice/MFCC.hpp
@@ -3,7 +3,7 @@
 # include <Siv3D.hpp>
 
 struct MFCC {
-	Array<float> feature; // 特徴量
+	Array<double> feature; // 特徴量
 
 	/// @brief 特徴量が埋まっていないかを調べる
 	/// @return 特徴量が埋まっていれば false
@@ -11,10 +11,10 @@ struct MFCC {
 
 	/// @brief ノルムを計算する
 	/// @return ノルム
-	[[nodiscard]] float norm() const;
+	[[nodiscard]] double norm() const;
 
 	/// @brief もう一つの MFCC とのコサイン類似度を計算する
 	/// @param other もう一つの MFCC
 	/// @return コサイン類似度
-	[[nodiscard]] float cosineSimilarity(const MFCC& other) const;
+	[[nodiscard]] double cosineSimilarity(const MFCC& other) const;
 };
